Navbar respons fixed
Changed it as a whole into a list which change direction depending on the device the customer will use, a mediaQuery is added for this function to work. One set back is that the cards in small device mode have a greater margin due to the navbar being fixed and gets higher cuz of the stack of the links.